% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affynorm.r
\name{affynorm}
\alias{affynorm}
\title{affynorm}
\usage{
affynorm(
  celfiles = NULL,
  customcdf = TRUE,
  versioncdf = NULL,
  norm = TRUE,
  bg = TRUE
)
}
\arguments{
\item{celfiles}{character vector contening cel files path}

\item{customcdf}{boolean TRUE by default use the last version of custom cdf. see details.}

\item{versioncdf}{character if you want to use old version of custom cdf. see details}

\item{norm}{logical apply quantile normalization}

\item{bg}{logical apply rma background correction}
}
\value{
data.frame of normalized data with features id as first column.
}
\description{
normalize affymetrix rawdata cel files using rma
and brainarry custom cdf
}
\details{
The function by default use customcdf = TRUE meaning that the it will install the last
version of custom cdf. You can choose an older version using for exemple versioncdf = 23.0.0.
customcdf = FALSE will use affymetrix probset design.
custom cdf version use ncbi entrez gene id.
}
\examples{
# not run
# affynorm(celfiles)
}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
