% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agilentrawdataextraction.r
\name{agilentrawdataextraction}
\alias{agilentrawdataextraction}
\title{extract rawdata for agilent array file}
\usage{
agilentrawdataextraction(files, output = TRUE, path = ".")
}
\arguments{
\item{files}{character rawdata file path}

\item{output}{logical}

\item{path}{character output direcotry}
}
\value{
lsit of 5 elements:
\itemize{
\item all data
\item rawdata matrix with control probes
\item rawdata matrix without control probes
\item matrix rawdata wihout control dimmension
\item annotations
}
}
\description{
extract rawdata for agilent array file
}
\details{
Extraction values from rawdata file
}
\examples{
# not run
# rawdatafilepath \%>\% rawdataextraction
}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
