% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agilentwf.r
\name{agilentwf}
\alias{agilentwf}
\title{agilent workflow for rawdata extraction , QCs and normalization}
\usage{
agilentwf(files, sif = NULL, bgfilter = 16, dirname = NULL, path = ".")
}
\arguments{
\item{files}{character list of rawdata files}

\item{sif}{data.frame sample meta data}

\item{bgfilter}{numeric threshold for background}

\item{dirname}{character}

\item{path}{character}
}
\value{
Create directory QC results
}
\description{
QC analysis for Agilent rawdata and normalisation for agilent chip rawdata
}
\details{
Results directory contain boxplot, acp, hierachical clustering for rawdata and dotplot of QC probes
A sample information file (sif) is created to complete with experimental factor variable.
Re run QCAgilent with competed sif for generated plot with that correspond to factor levels.
}
\examples{
# not run
# pathfile \%>\% qcagilent
}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
