% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.r
\name{biomart}
\alias{biomart}
\title{Biomart wrapper.}
\usage{
biomart(species = NULL)
}
\arguments{
\item{species}{character}
}
\value{
data.frame
}
\description{
Biomart wrapper.
}
\examples{
# not run
}
\author{
Florent Dumont \href{mailto:florent.dumont@univresite-paris-saclay.fr}{florent.dumont@univresite-paris-saclay.fr}
}
