% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customcdfInstall.r
\name{customcdfInstall}
\alias{customcdfInstall}
\title{install custom cdf}
\usage{
customcdfInstall(
  arraytype = NULL,
  version = NULL,
  annotationsource = "entrezg"
)
}
\arguments{
\item{arraytype}{character came from affynorm function}

\item{version}{character custom cdf version to install}

\item{annotationsource}{character}
}
\value{
character package name
}
\description{
Function to install custom cdf from brainarray site.
}
\details{
install custom cdf from brainarray for affymetrix .cel file normalization
}
\examples{
# not run
# celfiles \%>\% customcdfInstall
}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
