% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoseriematrix.r
\name{geoseriematrix}
\alias{geoseriematrix}
\title{Principal Component Analysis}
\usage{
geoseriematrix(file)
}
\arguments{
\item{file}{character path}
}
\value{
files
}
\description{
Principal Component Analysis
}
\examples{
# not run
# geoseriematrix( file )
}
\author{
Florent Dumont \href{mailto:florent.dumont@univresite-paris-saclay.fr}{florent.dumont@univresite-paris-saclay.fr}
}
