% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.r
\name{gff}
\alias{gff}
\title{Ensembl parser for gene, transcript, exon and prot.}
\usage{
gff(gff = NULL, pepfa = NULL)
}
\arguments{
\item{gff}{character path for gff ensembl file}

\item{pepfa}{character path for protein ensembl file}
}
\value{
data.frame
}
\description{
ggf file parser
}
\examples{
# not run
}
\author{
Florent Dumont \href{mailto:florent.dumont@univresite-paris-saclay.fr}{florent.dumont@univresite-paris-saclay.fr}
}
