% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepdt.r
\name{grepdt}
\alias{grepdt}
\title{grep in a data.frame}
\usage{
grepdt(pattern, dt, value = FALSE)
}
\arguments{
\item{pattern}{character to search in the data.frame}

\item{dt}{data.frame}

\item{value}{logical if FALSE return the data.frame of matching raws}
}
\value{
numeric vector of matching raws or values
}
\description{
Function that grep in a data.frame for each column
}
\details{
Function that make grep for a data.frame.
}
\examples{
# not run
# grepdt("a", dt , .value = TRUE )
}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
