% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipabarplotcp.r
\name{ipabarplotcp}
\alias{ipabarplotcp}
\title{ipabarplotcp}
\usage{
ipabarplotcp(dat, top = 80, labsize = 5)
}
\arguments{
\item{dat}{data.frame enrichment analysis results with 3 column}

\item{top}{numeric top feature to display}

\item{labsize}{numeric feature size}
}
\value{
plot
}
\description{
parse and create plot from ipa all enrichment results
}
\examples{
# not run
# ipabarplot( dat )
}
\author{
Florent Dumont \href{mailto:florent.dumont@univresite-paris-saclay.fr}{florent.dumont@univresite-paris-saclay.fr}
}
