% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orbitrapwf1.r
\name{orbitrapwf1}
\alias{orbitrapwf1}
\title{To extract data from spectrometry mzxml file}
\usage{
orbitrapwf1(files, .path = ".", .dirname = "orbitrapwf1")
}
\arguments{
\item{.path}{character}

\item{.dirname}{character}

\item{.files}{character vector of mzxml file}
}
\value{
data.frame
}
\description{
To extract data from spectrometry mzxml file
}
\examples{
# not run
# xcmsmzxml( listfiles )
}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
