% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pplipfilter.r
\name{pplipfilter}
\alias{pplipfilter}
\title{filter for phospholipid class}
\usage{
pplipfilter(dat, rtinfmin = NULL, rtsupmin = NULL, dirname = NULL, path = ".")
}
\arguments{
\item{dat}{matrix numeric}

\item{factor}{factor}

\item{samplename}{character}
}
\value{
results directory
}
\description{
filter for phospholipid class
}
\examples{
# not run
# pplipfilter(dat)
}
\author{
Florent Dumont \href{mailto:florent.dumont@univresite-paris-saclay.fr}{florent.dumont@univresite-paris-saclay.fr}
}
