% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcagilent.r
\name{qcagilent}
\alias{qcagilent}
\title{@title QC analysis for Agilent rawdata}
\usage{
qcagilent(files, sif = NULL, path = ".")
}
\arguments{
\item{files}{character list of rawdata files}

\item{sif}{data.frame with sample information file}

\item{path}{character}
}
\value{
Create directory QC results
}
\description{
QC analysis for Agilent rawdata expression array.
}
\details{
Results directory contain boxplot, acp, hierachical clustering for rawdata and dotplot of QC probes
A sample information file (sif) is created to complete with experimental factor variable.
Re run QCAgilent with competed sif for generated plot with that correspond to factor levels.
}
\examples{
# not run
# pathfile \%>\% qcagilent
}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
