% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacemat.r
\name{replacemat}
\alias{replacemat}
\title{replace value in data.frame}
\usage{
replacemat(dat, comptype = "<=", compvalue = 0, replacevalue = 1, dopar = TRUE)
}
\arguments{
\item{dat}{data.frame}

\item{comptype}{character for comparison operator}

\item{compvalue}{numeric or character value}

\item{replacevalue}{numeric or character value for replacement}

\item{dopar}{logical to use parallel computing}
}
\value{
data.frame
}
\description{
replace value in data.frame
}
\examples{
# not run
# replacemat(.dat)
}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
