% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpm.r
\name{rpm}
\alias{rpm}
\title{read per millions}
\usage{
rpm(dat)
}
\arguments{
\item{dat}{data.frame raw data count with row ID}
}
\value{
data.frame
}
\description{
Read per millions calculation from RNASeq raw data count.
}
\examples{
# not run
# rpm( dat )
}
\author{
Florent Dumont \href{mailto:florent.dumont@univresite-paris-saclay.fr}{florent.dumont@univresite-paris-saclay.fr}
}
