% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpm.r
\name{tpm}
\alias{tpm}
\title{Transcripts per millions}
\usage{
tpm(dat, length = NULL)
}
\arguments{
\item{dat}{matrix numeric}

\item{length}{numeric vector of base pair length}
}
\value{
data.frame
}
\description{
Read per millions calculation from RNASeq raw data count.
}
\examples{
# not run
# tpm( mat , length )
}
\author{
Florent Dumont \href{mailto:florent.dumont@univresite-paris-saclay.fr}{florent.dumont@univresite-paris-saclay.fr}
}
